Extract some tests to other files
