+++ test/query_generator/query_generator_test.dart
@@ -958,239 +958,6 @@ class SomeQuery with EquatableMixin {
       });
     });
 
-    test(
-        'If addQueryPrefix is true, all generated classes will have queryName as prefix',
-        () async {
-      final GraphQLQueryBuilder anotherBuilder =
-          graphQLQueryBuilder(BuilderOptions({
-        'generate_helpers': false,
-        'schema_mapping': [
-          {
-            'schema': 'api.schema.json',
-            'queries_glob': '**.graphql',
-            'output': 'lib/some_query.dart',
-            'add_query_prefix': true,
-          }
-        ]
-      }));
-      final GraphQLSchema schema = GraphQLSchema(
-          queryType: GraphQLType(name: 'Query', kind: GraphQLTypeKind.OBJECT),
-          types: [
-            GraphQLType(name: 'String', kind: GraphQLTypeKind.SCALAR),
-            GraphQLType(
-                name: 'MyEnum',
-                kind: GraphQLTypeKind.ENUM,
-                enumValues: [
-                  GraphQLEnumValue(name: 'value1'),
-                  GraphQLEnumValue(name: 'value2'),
-                ]),
-            GraphQLType(name: 'Query', kind: GraphQLTypeKind.OBJECT, fields: [
-              GraphQLField(
-                  name: 's',
-                  type: GraphQLType(
-                      name: 'String', kind: GraphQLTypeKind.SCALAR)),
-              GraphQLField(
-                  name: 'o',
-                  type: GraphQLType(
-                      name: 'SomeObject', kind: GraphQLTypeKind.OBJECT)),
-              GraphQLField(
-                  name: 'e',
-                  type:
-                      GraphQLType(name: 'MyEnum', kind: GraphQLTypeKind.ENUM)),
-            ]),
-            GraphQLType(
-                name: 'SomeObject',
-                kind: GraphQLTypeKind.OBJECT,
-                interfaces: [
-                  GraphQLType(
-                      name: 'AInterface', kind: GraphQLTypeKind.INTERFACE),
-                ],
-                fields: [
-                  GraphQLField(
-                      name: 'st',
-                      type: GraphQLType(
-                          name: 'String', kind: GraphQLTypeKind.SCALAR)),
-                  GraphQLField(
-                      name: 'ob',
-                      type: GraphQLType(
-                          kind: GraphQLTypeKind.LIST,
-                          ofType: GraphQLType(
-                              name: 'AnotherObject',
-                              kind: GraphQLTypeKind.OBJECT))),
-                ]),
-            GraphQLType(
-                name: 'AInterface',
-                kind: GraphQLTypeKind.INTERFACE,
-                fields: [
-                  GraphQLField(
-                      name: 'st',
-                      type: GraphQLType(
-                          name: 'String', kind: GraphQLTypeKind.SCALAR)),
-                ],
-                possibleTypes: [
-                  GraphQLType(name: 'SomeObject', kind: GraphQLTypeKind.OBJECT)
-                ]),
-            GraphQLType(
-                name: 'AnotherObject',
-                kind: GraphQLTypeKind.OBJECT,
-                fields: [
-                  GraphQLField(
-                      name: 'str',
-                      type: GraphQLType(
-                          name: 'String', kind: GraphQLTypeKind.SCALAR)),
-                ]),
-          ]);
-
-      final document = '''
-        query some_query {
-          s
-          o {
-            st
-            ob {
-              str
-            }
-          }
-          e
-        }
-        ''';
-
-      anotherBuilder.onBuild = expectAsync1((definition) {
-        expect(
-          definition,
-          LibraryDefinition(
-            'some_query',
-            queries: [
-              QueryDefinition(
-                'some_query',
-                parseString(document),
-                classes: [
-                  ClassDefinition(
-                      'SomeQuery',
-                      [
-                        ClassProperty('String', 's'),
-                        ClassProperty('SomeQuerySomeObject', 'o'),
-                        ClassProperty('SomeQueryMyEnum', 'e')
-                      ],
-                      prefix: 'SomeQuery'),
-                  ClassDefinition(
-                      'SomeQuerySomeObject',
-                      [
-                        ClassProperty('String', 'st', isOverride: true),
-                        ClassProperty('List<SomeQueryAnotherObject>', 'ob'),
-                        ClassProperty('String', 'resolveType',
-                            annotation: 'JsonKey(name: \'__resolveType\')',
-                            isOverride: true)
-                      ],
-                      implementations: ['SomeQueryAInterface'],
-                      prefix: 'SomeQuery'),
-                  ClassDefinition('SomeQueryAnotherObject',
-                      [ClassProperty('String', 'str')],
-                      prefix: 'SomeQuery'),
-                  ClassDefinition(
-                      'SomeQueryAInterface',
-                      [
-                        ClassProperty('String', 'st'),
-                        ClassProperty('String', 'resolveType',
-                            annotation: 'JsonKey(name: \'__resolveType\')')
-                      ],
-                      prefix: 'SomeQuery'),
-                  EnumDefinition('SomeQueryMyEnum', ['value1', 'value2']),
-                ],
-              ),
-            ],
-          ),
-        );
-      }, count: 1);
-
-      await testBuilder(anotherBuilder, {
-        'a|api.schema.json': jsonFromSchema(schema),
-        'a|some_query.graphql': document,
-      }, outputs: {
-        'a|lib/some_query.dart': '''// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:json_annotation/json_annotation.dart';
-import 'package:equatable/equatable.dart';
-import 'package:gql/ast.dart';
-part 'some_query.g.dart';
-
-@JsonSerializable(explicitToJson: true)
-class SomeQuery with EquatableMixin {
-  SomeQuery();
-
-  factory SomeQuery.fromJson(Map<String, dynamic> json) =>
-      _\$SomeQueryFromJson(json);
-
-  String s;
-
-  SomeQuerySomeObject o;
-
-  SomeQueryMyEnum e;
-
-  @override
-  List<Object> get props => [s, o, e];
-  Map<String, dynamic> toJson() => _\$SomeQueryToJson(this);
-}
-
-@JsonSerializable(explicitToJson: true)
-class SomeQuerySomeObject with EquatableMixin implements SomeQueryAInterface {
-  SomeQuerySomeObject();
-
-  factory SomeQuerySomeObject.fromJson(Map<String, dynamic> json) =>
-      _\$SomeQuerySomeObjectFromJson(json);
-
-  @override
-  String st;
-
-  List<SomeQueryAnotherObject> ob;
-
-  @override
-  @JsonKey(name: '__resolveType')
-  String resolveType;
-
-  @override
-  List<Object> get props => [st, ob, resolveType];
-  Map<String, dynamic> toJson() => _\$SomeQuerySomeObjectToJson(this);
-}
-
-@JsonSerializable(explicitToJson: true)
-class SomeQueryAnotherObject with EquatableMixin {
-  SomeQueryAnotherObject();
-
-  factory SomeQueryAnotherObject.fromJson(Map<String, dynamic> json) =>
-      _\$SomeQueryAnotherObjectFromJson(json);
-
-  String str;
-
-  @override
-  List<Object> get props => [str];
-  Map<String, dynamic> toJson() => _\$SomeQueryAnotherObjectToJson(this);
-}
-
-@JsonSerializable(explicitToJson: true)
-class SomeQueryAInterface with EquatableMixin {
-  SomeQueryAInterface();
-
-  factory SomeQueryAInterface.fromJson(Map<String, dynamic> json) =>
-      _\$SomeQueryAInterfaceFromJson(json);
-
-  String st;
-
-  @JsonKey(name: '__resolveType')
-  String resolveType;
-
-  @override
-  List<Object> get props => [st, resolveType];
-  Map<String, dynamic> toJson() => _\$SomeQueryAInterfaceToJson(this);
-}
-
-enum SomeQueryMyEnum {
-  value1,
-  value2,
-}
-''',
-      });
-    });
-
     test('Query name (pascal casing)', () async {
       final GraphQLQueryBuilder anotherBuilder =
           graphQLQueryBuilder(BuilderOptions({
@@ -1263,277 +1030,6 @@ class PascalCasingQuery with EquatableMixin {
   @override
   List<Object> get props => [s];
   Map<String, dynamic> toJson() => _\$PascalCasingQueryToJson(this);
-}
-''',
-      });
-    });
-
-    test('Fragments will have their own classes', () async {
-      final GraphQLQueryBuilder anotherBuilder =
-          graphQLQueryBuilder(BuilderOptions({
-        'generate_helpers': false,
-        'schema_mapping': [
-          {
-            'schema': 'api.schema.json',
-            'queries_glob': '**.graphql',
-            'output': 'lib/some_query.dart',
-          }
-        ]
-      }));
-      final GraphQLSchema schema = GraphQLSchema(
-          queryType:
-              GraphQLType(name: 'SomeObject', kind: GraphQLTypeKind.OBJECT),
-          types: [
-            GraphQLType(name: 'String', kind: GraphQLTypeKind.SCALAR),
-            GraphQLType(name: 'Int', kind: GraphQLTypeKind.SCALAR),
-            GraphQLType(
-                name: 'SomeObject',
-                kind: GraphQLTypeKind.OBJECT,
-                fields: [
-                  GraphQLField(
-                      name: 's',
-                      type: GraphQLType(
-                          name: 'String', kind: GraphQLTypeKind.SCALAR)),
-                  GraphQLField(
-                      name: 'i',
-                      type: GraphQLType(
-                          name: 'Int', kind: GraphQLTypeKind.SCALAR)),
-                ]),
-          ]);
-
-      anotherBuilder.onBuild = expectAsync1((definition) {
-        expect(
-          definition,
-          LibraryDefinition(
-            'some_query',
-            queries: [
-              QueryDefinition(
-                'some_query',
-                parseString(
-                    'fragment myFragment on SomeQuery { s, i }\nquery some_query { ...myFragment }'),
-                classes: [
-                  ClassDefinition('SomeQuery', [], mixins: [
-                    FragmentClassDefinition('MyFragmentMixin', [
-                      ClassProperty('String', 's'),
-                      ClassProperty('int', 'i'),
-                    ]),
-                  ]),
-                  FragmentClassDefinition('MyFragmentMixin', [
-                    ClassProperty('String', 's'),
-                    ClassProperty('int', 'i'),
-                  ]),
-                ],
-              ),
-            ],
-          ),
-        );
-      }, count: 1);
-
-      await testBuilder(anotherBuilder, {
-        'a|api.schema.json': jsonFromSchema(schema),
-        'a|some_query.query.graphql':
-            'fragment myFragment on SomeQuery { s, i }\nquery some_query { ...myFragment }',
-      }, outputs: {
-        'a|lib/some_query.dart': '''// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:json_annotation/json_annotation.dart';
-import 'package:equatable/equatable.dart';
-import 'package:gql/ast.dart';
-part 'some_query.g.dart';
-
-mixin MyFragmentMixin {
-  String s;
-  int i;
-}
-
-@JsonSerializable(explicitToJson: true)
-class SomeQuery with EquatableMixin, MyFragmentMixin {
-  SomeQuery();
-
-  factory SomeQuery.fromJson(Map<String, dynamic> json) =>
-      _\$SomeQueryFromJson(json);
-
-  @override
-  List<Object> get props => [s, i];
-  Map<String, dynamic> toJson() => _\$SomeQueryToJson(this);
-}
-''',
-      });
-    });
-
-    test('On union types', () async {
-      final GraphQLQueryBuilder anotherBuilder =
-          graphQLQueryBuilder(BuilderOptions({
-        'generate_helpers': false,
-        'schema_mapping': [
-          {
-            'schema': 'api.schema.json',
-            'queries_glob': '**.graphql',
-            'output': 'lib/some_query.dart',
-          }
-        ]
-      }));
-      final GraphQLSchema schema = GraphQLSchema(
-          queryType:
-              GraphQLType(name: 'SomeObject', kind: GraphQLTypeKind.OBJECT),
-          types: [
-            GraphQLType(name: 'Int', kind: GraphQLTypeKind.SCALAR),
-            GraphQLType(
-              name: 'SomeObject',
-              kind: GraphQLTypeKind.OBJECT,
-              fields: [
-                GraphQLField(
-                    name: 'o',
-                    type: GraphQLType(
-                        name: 'SomeUnion', kind: GraphQLTypeKind.UNION)),
-              ],
-            ),
-            GraphQLType(
-              name: 'TypeA',
-              kind: GraphQLTypeKind.OBJECT,
-              fields: [
-                GraphQLField(
-                    name: 'a',
-                    type:
-                        GraphQLType(name: 'Int', kind: GraphQLTypeKind.SCALAR)),
-              ],
-            ),
-            GraphQLType(
-              name: 'TypeB',
-              kind: GraphQLTypeKind.OBJECT,
-              fields: [
-                GraphQLField(
-                    name: 'b',
-                    type:
-                        GraphQLType(name: 'Int', kind: GraphQLTypeKind.SCALAR)),
-              ],
-            ),
-            GraphQLType(
-              name: 'SomeUnion',
-              kind: GraphQLTypeKind.UNION,
-              possibleTypes: [
-                GraphQLType(name: 'TypeA', kind: GraphQLTypeKind.OBJECT),
-                GraphQLType(name: 'TypeB', kind: GraphQLTypeKind.OBJECT),
-              ],
-            ),
-          ]);
-
-      anotherBuilder.onBuild = expectAsync1((definition) {
-        expect(
-          definition,
-          LibraryDefinition(
-            'some_query',
-            queries: [
-              QueryDefinition(
-                'some_query',
-                parseString(
-                    'query some_query { o { __typename, ... on TypeA { a }, ... on TypeB { b } } }'),
-                classes: [
-                  ClassDefinition('SomeQuery', [
-                    ClassProperty('SomeUnion', 'o'),
-                  ]),
-                  ClassDefinition('SomeUnion', [
-                    ClassProperty('String', 'resolveType',
-                        isOverride: true,
-                        annotation: 'JsonKey(name: \'__resolveType\')'),
-                  ], factoryPossibilities: [
-                    'TypeA',
-                    'TypeB'
-                  ]),
-                  ClassDefinition('TypeA', [ClassProperty('int', 'a')],
-                      extension: 'SomeUnion'),
-                  ClassDefinition('TypeB', [ClassProperty('int', 'b')],
-                      extension: 'SomeUnion'),
-                ],
-              ),
-            ],
-          ),
-        );
-      }, count: 1);
-
-      await testBuilder(anotherBuilder, {
-        'a|api.schema.json': jsonFromSchema(schema),
-        'a|some_query.query.graphql':
-            'query some_query { o { __typename, ... on TypeA { a }, ... on TypeB { b } } }',
-      }, outputs: {
-        'a|lib/some_query.dart': '''// GENERATED CODE - DO NOT MODIFY BY HAND
-
-import 'package:json_annotation/json_annotation.dart';
-import 'package:equatable/equatable.dart';
-import 'package:gql/ast.dart';
-part 'some_query.g.dart';
-
-@JsonSerializable(explicitToJson: true)
-class SomeQuery with EquatableMixin {
-  SomeQuery();
-
-  factory SomeQuery.fromJson(Map<String, dynamic> json) =>
-      _\$SomeQueryFromJson(json);
-
-  SomeUnion o;
-
-  @override
-  List<Object> get props => [o];
-  Map<String, dynamic> toJson() => _\$SomeQueryToJson(this);
-}
-
-@JsonSerializable(explicitToJson: true)
-class SomeUnion with EquatableMixin {
-  SomeUnion();
-
-  factory SomeUnion.fromJson(Map<String, dynamic> json) {
-    switch (json['__resolveType'].toString()) {
-      case 'TypeA':
-        return TypeA.fromJson(json);
-      case 'TypeB':
-        return TypeB.fromJson(json);
-      default:
-    }
-    return _\$SomeUnionFromJson(json);
-  }
-
-  @override
-  @JsonKey(name: '__resolveType')
-  String resolveType;
-
-  @override
-  List<Object> get props => [resolveType];
-  Map<String, dynamic> toJson() {
-    switch (resolveType) {
-      case 'TypeA':
-        return (this as TypeA).toJson();
-      case 'TypeB':
-        return (this as TypeB).toJson();
-      default:
-    }
-    return _\$SomeUnionToJson(this);
-  }
-}
-
-@JsonSerializable(explicitToJson: true)
-class TypeA extends SomeUnion with EquatableMixin {
-  TypeA();
-
-  factory TypeA.fromJson(Map<String, dynamic> json) => _\$TypeAFromJson(json);
-
-  int a;
-
-  @override
-  List<Object> get props => [a];
-  Map<String, dynamic> toJson() => _\$TypeAToJson(this);
-}
-
-@JsonSerializable(explicitToJson: true)
-class TypeB extends SomeUnion with EquatableMixin {
-  TypeB();
-
-  factory TypeB.fromJson(Map<String, dynamic> json) => _\$TypeBFromJson(json);
-
-  int b;
-
-  @override
-  List<Object> get props => [b];
-  Map<String, dynamic> toJson() => _\$TypeBToJson(this);
 }
 ''',
       });
